refactor: consulting-room에서 사용하는 컴포넌트들 파일 구조 정리
