+++ web/src/app/ui/consulting-room/info-content.tsx
