+++ web/src/app/ui/consulting-room/teller-consulting-room-side-bar.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import TabMenu from "./tab-menu";
-import Chatting from "./chatting";
+import Chatting from "../consulting-room/chatting";
 import Image from "next/image";
 import Logo from "@/app/utils/public/Logo.png";
 
@@ -10,7 +10,6 @@ export default function SideBar() {
       <TabMenu />
       <Chatting />
       <div className="flex justify-end">
-        {/* 로고 */}
         <Image src={Logo} alt="Logo" width={80} height={60} />
       </div>
     </div>
