+++ web/src/app/ui/consulting-room/waiting-content.tsx
